@HyukjinKwon 2022-06-01
This PR is ready for a look now. TL;DR:
- I added one new configuration `spark.sql.execution.arrow.localRelationThreshold` as a threshold.
- I fixed the code path to respect `spark.sql.execution.arrow.maxRecordsPerBatch` when creating Arrow batches to avoid forcing users to set a larger `spark.rpc.message.maxSize`.