@HyukjinKwon 2022-06-03
Gentle ping for a review :-). I know it has some trade-off but I believe this addresses more common cases and benefit more users.