@HyukjinKwon 2022-06-13
Let me get this in. It's the early stage of Spark 3.4 so should be good timing to merge such stuff.

Merged to master.